Moved Channels adapters to fragments.home_screen.adapters
