+++ app/src/main/java/com/tolstoy/zurichat/models/Channel.kt
@@ -1,0 +1,4 @@
+package com.tolstoy.zurichat.models
+
+//Dummy Data for populating the recyclerView
+data class Channel(var name: String, var privacy: Boolean, var read: Boolean )
